Helios 2.90 Prototype 0
This commit merges the HeFC Prototype 0 code committed in November
with Helios::TS 2.90 Prototype 0 code committed yesterday.  In order
to make the merged code functional, a few more modifications were required:

* Changes to Helios::Service to use the new Helios::Logger::Default instead
of the old Helios::Logger::Internal.

* Modified helios_job_info and helios_job_status commands to work with the
new HeFC Helios::Job API.

* A brand new helios_job_submit command based on the submit command in
HeliosX::Job::JSON.  In addition to working with the new HeFC Helios::Job
class, this new submit command also provides a cleaner command line syntax and
supports job arguments in JSON as well as XML syntax.

* Removed the old Makefile.PL and replaced it with HeFC_Makefile.PL from
the HeFC branch so the new modules and commands can be built properly.

While this code has been proven to process jobs under certain circumstances,
it is pre-alpha quality code and should not be used in any production
environment.  Additionally, the HeFC APIs are still very preliminary and
subject to change.  Given this early stage of development, we will not be
releasing a build of this code to CPAN.  Once development is farther along and
the new code and APIs are more stable, we we start to release alpha-quality
builds to CPAN.