New commands to work with jobtypes and queued jobs: * helios_job_lock * helios_job_unlock * helios_job_cancel * helios_jobtype_list
Each of the helios_job_* commands takes a jobid as a parameter,
and either locks the job in the queue, unlocks the job, or cancels
the job without processing it further.

The helios_jobtype_list command lists all of the jobtypes available
in the Helios collective database.

New command line shortcuts for helios_config_* commands.
Now, the user can skip switches and specify options positionally,
e.g. 'helios_config_get MyService myparam' will return the value
of the myparam parameter for MyService on the current host.
(This is just a shortcut; specifying command line switches still
works as before.